+++ neo/idlib/hashing/CRC32.h
@@ -11,8 +11,8 @@
 */
 
 void CRC32_InitChecksum( unsigned long &crcvalue );
-void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, int length );
+void CRC32_UpdateChecksum( unsigned long &crcvalue, const void *data, size_t length );
 void CRC32_FinishChecksum( unsigned long &crcvalue );
-unsigned long CRC32_BlockChecksum( const void *data, int length );
+unsigned long CRC32_BlockChecksum( const void *data, size_t length );
 
 #endif /* !__CRC32_H__ */
